Налоговые изменения

+++ code/modules/roguetown/roguemachine/ATM.dm
@@ -21,6 +21,7 @@
 		return
 
 	if(H in SStreasury.bank_accounts)
+		if(!skip_taxes(user)) return // REDMOON ADD - tax_redmoon - проверка налогов ИЛИ снятие денег для тех, кто их не оплачивает
 		var/amt = SStreasury.bank_accounts[H]
 		if(!amt)
 			say("Your balance is nothing.")
@@ -87,10 +88,14 @@
 		if(istype(P, /obj/item/roguecoin))
 			var/mob/living/carbon/human/H = user
 			if(H in SStreasury.bank_accounts)
+				var/obj/item/roguecoin/mammons = P
+				if(mammons.quantity > SStreasury.bank_accounts[H])
+					to_chat(user, span_info("I would overpay my taxes with it."))
+					return
 				SStreasury.generate_money_account(P.get_real_price(), H)
-				if(!(H.job in GLOB.noble_positions) && !HAS_TRAIT(H, TRAIT_NOBLE))
-					var/T = round(P.get_real_price() * SStreasury.tax_value)
-					say("Your deposit was taxed [T] mammon.")
+				//if(!(H.job in GLOB.noble_positions) && !HAS_TRAIT(H, TRAIT_NOBLE))
+					//var/T = round(P.get_real_price() * SStreasury.tax_value)
+					//say("Your deposit was taxed [T] mammon.")
 				qdel(P)
 				playsound(src, 'sound/misc/coininsert.ogg', 100, FALSE, -1)
 				return
@@ -100,5 +105,5 @@
 
 /obj/structure/roguemachine/atm/examine(mob/user)
 	. += ..()
-	. += span_info("The current tax rate on deposits is [SStreasury.tax_value * 100] percent. Nobles exempt.")
+	. += span_info("The current tax is [SStreasury.tax_value * 100] from the crown and [SStreasury.church_tith_value * 100] from the church.")
